Break up `LanguageConfigurationRegistryImpl` into multiple files
